+++ core-ng/src/main/java/core/framework/internal/web/service/WebServiceControllerBuilder.java
@@ -5,9 +5,12 @@ import core.framework.api.web.service.ResponseStatus;
 import core.framework.internal.asm.CodeBuilder;
 import core.framework.internal.asm.DynamicInstanceBuilder;
 import core.framework.internal.reflect.Params;
+import core.framework.log.Markers;
 import core.framework.web.Controller;
 import core.framework.web.Request;
 import core.framework.web.Response;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
 
 import java.lang.annotation.Annotation;
 import java.lang.reflect.Method;
@@ -36,13 +39,21 @@ public class WebServiceControllerBuilder<T> {
     public Controller build() {
         builder.addField("private final {} delegate;", type(serviceInterface));
         builder.constructor(new Class<?>[]{serviceInterface}, "this.delegate = $1;");
-        builder.addMethod(buildMethod());
+
+        boolean deprecated = method.isAnnotationPresent(Deprecated.class);
+        if (deprecated) {
+            builder.addField("private final {} logger = {}.getLogger({});", type(Logger.class), type(LoggerFactory.class), variable(serviceInterface));
+        }
+        builder.addMethod(buildMethod(deprecated));
+
         return builder.build(service);
     }
 
-    private String buildMethod() {
+    private String buildMethod(boolean deprecated) {
         var builder = new CodeBuilder();
         builder.append("public {} execute({} request) throws Exception {\n", type(Response.class), type(Request.class));
+
+        if (deprecated) builder.indent(1).append("logger.warn({}.errorCode({}), {});\n", type(Markers.class), variable("DEPRECATION"), variable("web service has been deprecated, please notify consumer to update"));
 
         Annotation[][] annotations = method.getParameterAnnotations();
         List<String> params = new ArrayList<>(annotations.length);