+++ ext/monitor/src/main/java/app/monitor/kafka/EventMessageHandler.java
@@ -24,6 +24,7 @@ public class EventMessageHandler implements MessageHandler<EventMessage> {
         var alert = new Alert();
         alert.id = message.id;
         alert.app = message.app;
+        alert.action = message.action;
         alert.severity(message.result);
         alert.errorCode = message.errorCode;
         alert.errorMessage = message.errorMessage;